Update main.py
[Issue] Bug of lr parameter type(int) error. [Solution] revise it as float32.

+++ src/main.py
@@ -39,7 +39,7 @@ parser.add_argument('--trainOrtest', type=str, default="train", help='status of 
 # training hyperparameters
 parser.add_argument('--optim', type=str, default="Adam", help='optimizer for training', choices=['RMSprop', 'SGD', 'Adam'])
 parser.add_argument('--batch_size', type=int, default=32, help='input batch size for training (default: 64)')
-parser.add_argument('--lr', type=int, default=1e-4, help='initial learning rate for training (default: 1e-3)')
+parser.add_argument('--lr', type=float, default=1e-4, help='initial learning rate for training (default: 1e-3)')
 parser.add_argument('--CosineAnnealingWarmRestarts', type=bool, default=False, help='initial learning rate for training (default: 1e-3)')
 parser.add_argument('--epochs', type=int, default=50, help='number of epochs to train (default: 10)')
 parser.add_argument('--grad_scale', type=float, default=8, help='learning rate for wage delta calculation')
